# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:3001/api","description":"Generated server url"}],"paths":{"/rentals/{id}":{"get":{"tags":["rental-controller"],"operationId":"getRentalById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RentalResponseDTO"}}}}}},"put":{"tags":["rental-controller"],"operationId":"updateRental","parameters":[{"name":"updatedData","in":"query","required":true,"schema":{"$ref":"#/components/schemas/RentalPutRequestDTO"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiMessageResponse"}}}}}}},"/rentals":{"get":{"tags":["rental-controller"],"operationId":"getRentals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiRentalResponse"}}}}}},"post":{"tags":["rental-controller"],"operationId":"createRental","parameters":[{"name":"rentalPostRequestDTO","in":"query","required":true,"schema":{"$ref":"#/components/schemas/RentalPostRequestDTO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiMessageResponse"}}}}}}},"/messages":{"post":{"tags":["message-controller"],"operationId":"createMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiMessageResponse"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiTokenResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiTokenResponse"}}}}}}},"/user/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/pictures/{fileName}":{"get":{"tags":["picture-controller"],"operationId":"getPicture","parameters":[{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/auth/me":{"get":{"tags":["auth-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}}},"components":{"schemas":{"RentalPutRequestDTO":{"required":["description","name","price","surface"],"type":"object","properties":{"name":{"type":"string"},"surface":{"type":"number","format":"double"},"price":{"type":"number","format":"double"},"description":{"type":"string"}}},"ApiMessageResponse":{"type":"object","properties":{"message":{"type":"string"}}},"RentalPostRequestDTO":{"required":["description","name","picture","price","surface"],"type":"object","properties":{"name":{"type":"string"},"surface":{"type":"number","format":"double"},"picture":{"type":"string","format":"binary"},"price":{"type":"number","format":"double"},"description":{"type":"string"}}},"MessageDTO":{"required":["message","rental_id","user_id"],"type":"object","properties":{"rental_id":{"type":"integer","format":"int64"},"user_id":{"type":"integer","format":"int64"},"message":{"type":"string"}}},"UserRequestDTO":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"ApiTokenResponse":{"type":"object","properties":{"token":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}},"ApiRentalResponse":{"type":"object","properties":{"rentals":{"type":"object"}}},"RentalResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"surface":{"type":"number","format":"double"},"price":{"type":"number","format":"double"},"picture":{"type":"string"},"description":{"type":"string"},"ownerId":{"type":"integer","format":"int64"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}}}}}