# src/main/java/com/chatop/api/service/DTO/apiResponse/ApiMessageResponse.java
package com.chatop.api.service.DTO.apiResponse;

import lombok.Data;

@Data
public class ApiMessageResponse {

	private String message;
	
    public ApiMessageResponse(String message) {
        this.message = message;
    }
}